Display activity results to main activity screen

+++ app/src/main/java/com/lovricante/zavrsnirad/ActivityData.java
@@ -6,11 +6,11 @@ import java.util.ArrayList;
 public class ActivityData implements Serializable {
     private String activityType;
     private long startTime;
-    private long duration;
+    private int duration;
     private float distance;
     private ArrayList<TimePlace> timePlaces;
 
-    public ActivityData(String activityType, long startTime, long duration, float distance, ArrayList<TimePlace> timePlaces) {
+    public ActivityData(String activityType, long startTime, int duration, float distance, ArrayList<TimePlace> timePlaces) {
         this.activityType = activityType;
         this.startTime = startTime;
         this.duration = duration;
